Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	renamed:    gen_ai_python/mayplot/_intro.py -> gen_ai_python/mayplot/1_intro.py
	renamed:    gen_ai_python/mayplot/line_plot.py -> gen_ai_python/mayplot/2_line_plot.py
	renamed:    gen_ai_python/mayplot/bar_plot.py -> gen_ai_python/mayplot/3_bar_plot.py
	renamed:    gen_ai_python/mayplot/pie_plot.py -> gen_ai_python/mayplot/4_pie_plot.py
	renamed:    gen_ai_python/mayplot/scatterr_plot.py -> gen_ai_python/mayplot/5_scatterr_plot.py
	renamed:    gen_ai_python/mayplot/histogram_plot.py -> gen_ai_python/mayplot/6_histogram_plot.py
	deleted:    gen_ai_python/mayplot/tempCodeRunnerFile.py
	modified:   gen_ai_python/numpy/1_intro.py

+++ gen_ai_python/numpy/1_intro.py
@@ -83,9 +83,10 @@ print("sum of rows in 2d array",np.sum(my_2d_array,axis=1))
 
 
 #creating special arrays
-zeros_array = np.zeros((3,4))
+zeros_array = np.zeros((3,4),dtype="int")
 print("3x4 array of zeros\n",zeros_array)
 ones_array = np.ones((2,3))
+
 print("2x3 array of ones\n",ones_array)
 full_array = np.full((2,2),7)
 print("2x2 array of sevens\n",full_array)
